pyproject.toml
[project]
name = "labor-dashboard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "duckdb>=1.5.0",
    "httpx>=0.28.1",
    "openai>=2.29.0",
    "polars>=1.39.0",
    "python-dotenv>=1.2.2",
    "yfinance>=1.2.0",
]
